pyproject.toml
[project]
name = "kafka-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python-ng>=2.2.3",
]
